Committing the first working version of reading in precomputed data of refractive index and displacements from a run of kWave. This uses a new class implemented (InputHDF5Stream) to read the HDF5 file.  This has been verified to read in data correctly into the C++ version by reading in data into matlab from the HDF5 file.
New commandline flags have been created to facilitate this.
Note that the input file is the file that generated the output file
(HDF5 files), and it's the output file that is read in and populates
the data that will be used in the MC-sim.

The dimensions of the 3D array have been changed to match what is
actually used during execution (i.e. Fulldim).  Reference
MatrixContainer.cpp for details.

To do:
- Move away from the test routine and have the data that is read
  in be handed off to the monte-carlo simulation.
- Verify that correct indexing is done into the 3-D array.
- What eventually needs to be done is verification that
  the values computed for the refractive index and displacement
  are sane.