Trying to figure out why modulation no longer happens.  As of now this commit is deemed broken by an unknown change that results in photons that don't interact with ultrasound to change their OPL. This means that something in the indexing in the displacement data (for example) is incorrect, or the data being handed off to the monte-carlo simulation doesn't contain the information I assume it does (i.e. displacements are zero).
This commit contains changes to Photon::initTrajectory(), which was improperly
setting the direction cosines of the photon, at the start of propagation, to
act as an isotropic source.  The direction now is fully along the optical
axis upon creation of the photon.  Also, the delta stepsize was being set
in cm due to the absorption and scattering coefficients being given in
units of cm^-1.  They have been multiplied to put them into m^-1; verify
that this is correct.

To do:
- Figure out why the OPL of photons are not changing when displacement
  is enabled. Is it a problem with indexing into the displacement data?
  Or maybe the displacement arrays are not containing values that are
  being updated each time step?
- Verify that the change in the step size during Photon::Hop() is
  correct.